checking for table existence before dropping

+++ db/migrate/20160527114007_drop_unnecessary_tables.rb
@@ -1,6 +1,6 @@
 class DropUnnecessaryTables < ActiveRecord::Migration
   def up
-    drop_table :organization_roles
-    drop_table :roles
+    drop_table :organization_roles if table_exists?(:organization_roles)
+    drop_table :roles if table_exists?(:roles)
   end
 end
